fix: expose CLI binary from public rust crate
