+++ crates/mcp-compressor-core/src/app/mod.rs
@@ -1,4 +1,5 @@
 pub mod banner;
+pub mod entrypoint;
 pub mod options;
 pub mod paths;
 pub mod runtime;
