+++ crates/mcp-compressor/src/main.rs
@@ -1,0 +1,5 @@
+use std::process::ExitCode;
+
+fn main() -> ExitCode {
+    mcp_compressor::app::entrypoint::main_exit_code()
+}
